Do final PR review comments
