+++ src/components/RequestPanel/RequestStatus.js
@@ -1,13 +1,10 @@
-import React, { useMemo } from 'react'
-import { Button, GU, IconCheck, IconCross, Info, useTheme } from '@aragon/ui'
-import RequestStepItem from './RequestStepItem'
+import React from 'react'
+import { Button, GU, Info } from '@aragon/ui'
+import RequestSteps from './Steps/RequestSteps'
 import { useWallet } from '../../providers/Wallet'
 import { getProviderFromUseWalletId } from '../../ethereum-providers'
 import {
-  REQUEST_STATUS_CONFIRMED,
   REQUEST_STATUS_FAILED,
-  REQUEST_STATUS_PENDING,
-  REQUEST_STATUS_PROCESSED,
   REQUEST_STATUS_PROCESSING_FAILED,
 } from './request-statuses'
 
@@ -20,32 +17,16 @@ function RequestStatus({
   requests,
   requestStatus,
 }) {
-  const theme = useTheme()
-
   const multipleRequests = requests.length > 1
 
   return (
     <div>
-      <div
-        css={`
-          background: ${theme.feedbackSurface};
-          height: ${44 * GU}px;
-          display: flex;
-          flex-direction: column;
-          justify-content: space-evenly;
-          align-items: center;
-        `}
-      >
-        {multipleRequests ? (
-          <MultiStepRequest
-            lastProcessedAt={lastProcessedAt}
-            requests={requests}
-            requestStatus={requestStatus}
-          />
-        ) : (
-          <SingleRequest request={requests[0]} status={requestStatus[0]} />
-        )}
-      </div>
+      <RequestSteps
+        lastProcessedAt={lastProcessedAt}
+        multipleRequests={multipleRequests}
+        requests={requests}
+        requestStatus={requestStatus}
+      />
       <div
         css={`
           margin-top: ${2 * GU}px;
@@ -65,109 +46,6 @@ function RequestStatus({
   )
 }
 
-function MultiStepRequest({ lastProcessedAt, requests, requestStatus }) {
-  return requests.map((request, index) => {
-    const status = requestStatus[index]
-
-    return (
-      <RequestStepItem
-        key={index}
-        lastProcessedAt={lastProcessedAt}
-        status={status}
-        stepNumber={index + 1}
-        request={request}
-      />
-    )
-  })
-}
-
-function SingleRequest({ request, status }) {
-  const theme = useTheme()
-
-  const { iconColor, label } = useMemo(() => {
-    if (status === REQUEST_STATUS_PROCESSING_FAILED) {
-      return {
-        iconColor: theme.negative,
-        label: request.isTx ? 'Signing transaction failed!' : request.onError,
-      }
-    }
-
-    if (status === REQUEST_STATUS_CONFIRMED) {
-      return {
-        iconColor: theme.positive,
-        label: 'Transaction confirmed!',
-      }
-    }
-
-    if (status === REQUEST_STATUS_PENDING) {
-      return {
-        iconColor: theme.info,
-        label: 'Transaction being mined…',
-      }
-    }
-
-    if (status === REQUEST_STATUS_PROCESSED) {
-      return {
-        iconColor: theme.positive,
-        label: request.isTx ? 'Transaction signed!' : request.onSuccess,
-      }
-    }
-
-    return {
-      iconColor: theme.hint,
-      label: request.isTx ? 'Waiting for signature…' : request.description,
-    }
-  }, [request, status, theme])
-
-  return (
-    <div
-      css={`
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-      `}
-    >
-      <div
-        css={`
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          width: ${7.5 * GU}px;
-          height: ${7.5 * GU}px;
-          border: 2px solid ${iconColor};
-          border-radius: 50%;
-          transition: border-color 150ms ease-in-out;
-        `}
-      >
-        {status === REQUEST_STATUS_PROCESSING_FAILED ? (
-          <IconCross
-            size="medium"
-            css={`
-              color: ${iconColor};
-              transition: color 150ms ease-in-out;
-            `}
-          />
-        ) : (
-          <IconCheck
-            size="medium"
-            css={`
-              color: ${iconColor};
-              transition: color 150ms ease-in-out;
-            `}
-          />
-        )}
-      </div>
-      <p
-        css={`
-          margin-top: ${3 * GU}px;
-        `}
-      >
-        {label}
-      </p>
-    </div>
-  )
-}
-
 function RequestStatusInfo({
   allSuccess,
   multipleRequests,
@@ -182,7 +60,7 @@ function RequestStatusInfo({
 
   if (allSuccess) {
     return (
-      <Info>Request{multipleRequests ? 's' : ''} processed successfully!</Info>
+      <Info>Step{multipleRequests ? 's' : ''} processed successfully!</Info>
     )
   }
 
@@ -207,8 +85,8 @@ function RequestStatusInfo({
       <>
         <Info>
           {requestFailedStatus === REQUEST_STATUS_PROCESSING_FAILED
-            ? "Your request wasn't processed correctly"
-            : 'Your request failed'}
+            ? "This step wasn't processed correctly"
+            : 'This step failed'}
         </Info>
         <Button
           css={`
@@ -228,15 +106,16 @@ function RequestStatusInfo({
       <Info>
         {requiresMultipleSignatures ? (
           <span>
-            This action requires multiple transactions. Please sign them one
-            after another and <b>do not close this window</b> until the process
-            is finished. Open your Ethereum provider {provider.name} to sign the
-            transactions.{' '}
+            These actions require multiple transactions. Please sign them one
+            after another and <strong>do not close this window</strong> until
+            all of them are finished. Open {provider.name} to sign the
+            transactions.
           </span>
         ) : (
           <span>
-            This action requires multiple steps. Please,{' '}
-            <b>do not close this window</b> until the process is finished.
+            This action involves multiple steps. Please,{' '}
+            <strong>do not close this window</strong> until all of them are
+            finished.
           </span>
         )}
       </Info>
@@ -247,7 +126,7 @@ function RequestStatusInfo({
   return (
     <Info>
       {requests[0].isTx
-        ? `Open ${provider.name} to sign your transaction`
+        ? `Open ${provider.name} to sign the transaction`
         : requests[0].description}
     </Info>
   )
@@ -262,9 +141,9 @@ const RequestReattempt = ({
     <>
       <Info>
         {maxAttemptsReached
-          ? `Seems that the request can't be processed`
-          : `An error has occurred during the request process. Don't worry, you can
-             try to process the request again.`}
+          ? `It seems there's something wrong with this step`
+          : `An error occurred at this step. Don't worry, you can
+             try to process this step again.`}
       </Info>
       <Button
         css={`
